# services/chart-service/api/openapi.yaml
openapi: 3.0.3
info:
  title: chart-service API
  version: "1.0.0"
  description: >
    Dashboards + charts for the Insights and Case Management modules (BRD 07).
    Resolves chart data by compiling chart definitions through semantic-service
    and executing through query-service, applies server-side aggregation by
    default, caches shaped results in Redis (ETag/304), and serves drilldowns
    and CSV/PNG exports. All endpoints require an RS256 JWT (MASTER-FR-010);
    authorization is via the local OPA sidecar (MASTER-FR-012).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /charts/{id}/export:
    post:
      summary: Async CSV/PNG export (CHART-FR-041)
      responses: { '202': { description: '{operation_id}' }, '429': { description: EXPORT_LIMIT } }
...
  /exports/{key}:
    get:
      summary: Download a signed export artifact (HMAC-validated, not JWT)
      parameters:
        - { name: key, in: path, required: true, schema: { type: string } }
        - { name: exp, in: query, required: true, schema: { type: integer } }
        - { name: sig, in: query, required: true, schema: { type: string } }
      responses: { '200': { description: artifact }, '404': { description: expired/invalid } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }